feat: Add table to cross check user migration status
This branch adds support to look in a `migration` table for a user's
`fxa_uid` record. If it is found, the server will return a 5xx error to
the client preventing sync operations.

NOTE: IT IS IMPORTANT THAT THE SERVER NOT RETURN A 2xx RESULT TO ANY
CLIENT IN MIGRATION STATUS UNLESS THE `meta`/`global` BSO IS RESET.

The `migration` table is created by the `user_migration` script, which
is part of `syncstorage-rs`.