Use `requirements.txt` as source of truth for requirements in setup.py
Previously, this repo had two sets of requirements: one in `requirements.txt`
that we'd use when building the project via `make`, and one in `setup.py`
that we'd use if installing it as a depenency via pip.

This has been causing problems for self-hosters, so let's have `setup.py`
just read the list of requirements out of `requirements.txt` and avoid
needing to keep two lists of requirements in sync.

(The logic here is copied directly from the tokenserver `setup.py`,
and I've taken the opportunity to copy over some other `Makefile` tweaks
while I'm here).